requirements.txt
django==3.1.4
djangorestframework-gis==0.16
gunicorn==20.0.4
python-dotenv==0.15.0
psycopg2-binary==2.8.6
GDAL==3.1.4
Pillow==8.0.1
django-admin-sortable2==0.7.7
django-tinymce==3.2.0
requests==2.25.1